+++ build_config/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 - The `target` and `build_extensions` keys for builder definitions are now
   optional and should be omitted in most cases since they are currently unused.
+- Support options based on mode, add `devOptions` and `releaseOptions` on
+  `TargetBuilderConfig`.
+- Support applying default options based on builder definitions, add `option`,
+  `devOptions`, and `releaseOptions` to `TargetBuilderConfigDefaults`.
+- Ensure that `defaults` and `generateFor` fields are never null.
+- Add `InputSet.anything` to name the input sets that don't filter out any
+  assets.
 
 ## 0.2.5